final-project/requirements.txt
cryptography == 42.0.5
Flask == 3.0.2
openai == 1.14.1
python-dotenv == 1.0.1
